hourly calculations done
