+++ main.py
@@ -17,7 +17,6 @@ from cspr_summarization.entities.PairSyncEvent import *
 from cspr_summarization.entities.TokenTotalSupply import *
 import pandas as pd
 from cspr_summarization.services.lp_hourly_summarizer import LpHourlySummarizer
-from cspr_summarization.services.lp_hourly_summarizer2 import LpHourlySummarizer2
 import pytz
 from datetime import datetime
 
@@ -33,9 +32,11 @@ timestamp_obj = datetime.strptime(timestamp_string, '%Y-%m-%d %H:%M:%S.%f %z')
 end_hour = timestamp_obj.astimezone(pytz.UTC)
 
 
-summarizer2 = LpHourlySummarizer2(start_hour, end_hour)
-summarizer2.init_hourly_data()
-summarizer2.sync_summarization()
-summarizer2.mint_summarization()
-summarizer2.burn_summarization()
-summarizer2.update_close_lp_token_supply()
+summarizer = LpHourlySummarizer(start_hour, end_hour)
+summarizer.init_hourly_data()
+summarizer.sync_summarization()
+summarizer.mint_summarization()
+summarizer.burn_summarization()
+summarizer.swap_summarization()
+summarizer.update_close_lp_token_supply()
+summarizer.set_max_block()
